feat: add v1 schema migration capability
