+++ schemas/v0/0_null.go
@@ -3,5 +3,5 @@ package v0
 func init() {
 	up := batch(`SELECT 1;`)
 	down := batch(`SELECT 1;`)
-	Patches.MustRegisterTx(up, down)
+	patches.MustRegisterTx(up, down)
 }
