+++ schemas/v0/12_derived_consensus_chain.go
@@ -37,5 +37,5 @@ WITH NO DATA;
 DROP MATERIALIZED VIEW IF EXISTS derived_consensus_chain_view;
 	`)
 
-	Patches.MustRegisterTx(up, down)
+	patches.MustRegisterTx(up, down)
 }
