+++ schemas/v0/22_visor_processing_reports.go
@@ -22,5 +22,5 @@ CREATE TABLE IF NOT EXISTS public.visor_processing_reports (
 	DROP TABLE IF EXISTS public.visor_processing_reports;
 `)
 
-	Patches.MustRegisterTx(up, down)
+	patches.MustRegisterTx(up, down)
 }
