+++ schemas/v0/23_miner_sector_deals_pk.go
@@ -37,5 +37,5 @@ func init() {
 		SELECT 1;
 `)
 
-	Patches.MustRegisterTx(up, down)
+	patches.MustRegisterTx(up, down)
 }
