+++ schemas/v0/24_height_functions_fast.go
@@ -22,5 +22,5 @@ func init() {
 	DROP FUNCTION IF EXISTS public.height_to_unix;
 `)
 
-	Patches.MustRegisterTx(up, down)
+	patches.MustRegisterTx(up, down)
 }
