+++ schemas/v0/29_miner_info_sector_size.go
@@ -11,5 +11,5 @@ func init() {
 		ALTER TABLE public.miner_infos DROP COLUMN sector_size;
 `)
 
-	Patches.MustRegisterTx(up, down)
+	patches.MustRegisterTx(up, down)
 }
